+++ src/slurmctld/reservation.c
@@ -711,7 +711,7 @@ static int _set_assoc_list(slurmctld_resv_t *resv_ptr)
 				memset(&assoc, 0,
 				       sizeof(slurmdb_assoc_rec_t));
 				assoc.acct = resv_ptr->account_list[j];
-				assoc.uid  = (uint32_t)NO_VAL;
+				assoc.uid  = NO_VAL;
 				rc = _append_assoc_list(assoc_list, &assoc);
 				if (rc != SLURM_SUCCESS)
 					goto end_it;
@@ -743,7 +743,7 @@ static int _set_assoc_list(slurmctld_resv_t *resv_ptr)
 		for (i=0; i < resv_ptr->account_cnt; i++) {
 			memset(&assoc, 0, sizeof(slurmdb_assoc_rec_t));
 			assoc.acct = resv_ptr->account_list[i];
-			assoc.uid  = (uint32_t)NO_VAL;
+			assoc.uid  = NO_VAL;
 			if ((rc = _append_assoc_list(assoc_list, &assoc))
 			    != SLURM_SUCCESS) {
 				goto end_it;
